# This is a combination of 12 commits. # This is the 1st commit message:
Add functionality to fix areas based on zip prefixes to US zip code fix command

# This is the commit message #2:

Fix flake8 errors

# This is the commit message #3:

Added update registered address area command

# This is the commit message #4:

Remove migration for postcodes and company states

# This is the commit message #5:

Developed tests and refactored code with tests

# This is the commit message #6:

Moved large file to constants

# This is the commit message #7:

Add new un-nested proposition endpoint

# This is the commit message #8:

Add latest interaction to investment projects es model

# This is the commit message #9:

Tests for latest interaction

# This is the commit message #10:

Deal with es sync for changing interaction's project

# This is the commit message #11:

Added changelog for latest interaction

# This is the commit message #12:

Prepare for release 36.9.0

This bumps the version and adds the changelog for version 36.9.0.
