--- changelog/area_code.db.md
@@ -1,1 +1,0 @@
-A new field `area_code` has been added to Administrative Areas under metadata.
