--- datahub/dbmaintenance/test/commands/test_fix_us_company_address_postcode_data.py
@@ -1,4 +1,0 @@
-"""
-    TODO: Do the integration test creating several scenarios of bad
-        ßß  data that this should fix
-"""
